+++ R/MarkovTransition.R
@@ -29,13 +29,12 @@ MarkovTransition <- R6::R6Class(
     #' @param target \code{MarkovState} to which the transition ends.
     #' @param r Instantaneous hazard rate in units of per patient per year.
     #' Exactly one of the outgoing transitions from each state must be NULL.
-    #' Default value is \code{NA} to indicate that rates will be set with
-    #' \code{set_rate} as the model is run.
+    #' Default value is NULL.
     #' @param cost Cost associated with the transition. 
     #' @param label Character string containing a label for the transition (the
     #' name of the event).
     #' @return A new \code{MarkovTransition} object.
-    initialize = function(source, target, r=as.numeric(NA), cost=0, label="") {
+    initialize = function(source, target, r=NULL, cost=0, label="") {
       # initialize base class
       super$initialize(source=source, target=target, label=label)
       # check that source inherits from MarkovState
@@ -92,11 +91,11 @@ MarkovTransition <- R6::R6Class(
 
     #' @description Set the value of the transition rate
     #' @param r Instantaneous hazard rate in units of per patient per year,
-    #' or NULL.
+    #' or NULL; default value is zero.
     #' @details Exactly one of the outgoing transitions from each state must
     #' be NULL.
-    #' @return Updated \code{MarkovTransition} cost
-    set_rate = function(r) {
+    #' @return Updated \code{MarkovTransition} object.
+    set_rate = function(r=0) {
       # check r and set private variable
       if (!is.null(r)) {
         if (inherits(r, what="numeric")) {
@@ -110,6 +109,7 @@ MarkovTransition <- R6::R6Class(
       } else {
         private$transition.rate <- NULL
       }
+      return(invisible(self))
     },
     
     #' @description Return the value of the hazard rate.